refactor(daemon): rename directories, namespaces and move command handlers
