+++ src/daemon/Tiempitod.NET/Commands/Configuration/DisableConfigCommand.cs
@@ -1,6 +1,6 @@
 using Tiempitod.NET.Configuration.User;
 
-namespace Tiempitod.NET.Commands.ConfigCommands;
+namespace Tiempitod.NET.Commands.Configuration;
 
 /// <summary>
 /// Represents the command to disable a user's feature configuration.
