+++ src/daemon/Tiempitod.NET/Commands/Configuration/EnableConfigCommand.cs
@@ -1,6 +1,6 @@
 using Tiempitod.NET.Configuration.User;
 
-namespace Tiempitod.NET.Commands.ConfigCommands;
+namespace Tiempitod.NET.Commands.Configuration;
 
 /// <summary>
 /// Represents the command to enable a user's feature configuration.
