+++ src/daemon/Tiempitod.NET/Commands/SessionManagement/CancelSessionCommand.cs
@@ -1,6 +1,6 @@
 using Tiempitod.NET.Session;
 
-namespace Tiempitod.NET.Commands.SessionCommands;
+namespace Tiempitod.NET.Commands.SessionManagement;
 
 /// <summary>
 /// Command that cancels the current active session.
