+++ src/daemon/Tiempitod.NET/Commands/SessionManagement/ResumeSessionCommand.cs
@@ -1,6 +1,6 @@
 using Tiempitod.NET.Session;
 
-namespace Tiempitod.NET.Commands.SessionCommands;
+namespace Tiempitod.NET.Commands.SessionManagement;
 
 /// <summary>
 /// Command that continues a paused session.
